% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{ForensicDocument-package}
\alias{ForensicDocument}
\alias{ForensicDocument-package}
\title{ForensicDocument package}
\description{
This package provides operational and useable methods and techniques for the analysis of handwriting, notably quantification of characters' shape and (ii) rigurous probabilistic inferntial procedures for coherent assesment of handwriting evidence. This package is developed in the School of Forensic Science - University of Lausanne.
}
\author{
Alexandre Thiery, Silvia Bozza
}

