% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ImageGetClosedShapeRadius}
\alias{ImageGetClosedShapeRadius}
\title{Extracts the radius of a closed shape, relative to its barycenter.}
\usage{
ImageGetClosedShapeRadius(image, seq)
}
\arguments{
\item{image}{A numeric matrix, representing the black and white image (a pixel of value \code{0} is from the background).}

\item{seq}{A numeric vector, giving the angles to subsample the closed shape.}
}
\value{
A numeric vector containing the radius values.
}
\description{
This method allows to extract  the radius of a closed shape, contained in a image, relative to its barycenter.
}
\author{
Alexandre Thiery
}
\keyword{internal}

