% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ImageRead}
\alias{ImageRead}
\title{Read a bitmap image into a binary matrix}
\usage{
ImageRead(file, character_pixel = 1)
}
\arguments{
\item{file}{A character: name of the file to read from.}

\item{character_pixel}{A interger value, indicating which pixel value is from the character.}
}
\description{
Reads an image stored from various file format (see Details) into a matrix. of \code{0}'s and \code{1}'s.
}
\details{
This method allows to read images of format type:
\enumerate{
  \item PNG, using the \pkg{png} package.
}
}
\author{
Alexandre Thiery
}
\seealso{
the method \code{readPNG} from the package \code{png}.
}
\keyword{internal}

