% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ImageThin}
\alias{ImageThin}
\title{Extract the skeleton of a black and white image.}
\usage{
ImageThin(matrix)
}
\arguments{
\item{matrix}{A numeric matrix, representing the black and white image (a pixel of value \code{0} is from the background).}
}
\value{
A matrix with values of \code{1} for the contour and values of \code{0} for the background.
}
\description{
Extracts the skeleton of a black and white image (that is, represented by a matrix).
}
\details{
The skeletonisation algorithm implemented is based on \emph{Stentiford (1983)}. It is a recursive thinning process that allows to extract a curve with a thickness of 1 pixel.
}
\examples{
\dontrun{
## We use the 'fig-O.png' example image available in this package
file = system.file("extdata", "fig-O.png", package = "ForensicDocument")

## Read the image using 'ImageRead' method, or (similarly with 'png::readPNG')
image1 = ForensicDocument:::ImageRead(file, character_pixel = 0)

## Create the image skeleton
image2 = ForensicDocument:::ImageThin(image1)

par(mfrow = c(1,2))
image(t(image1), xlim = c(0,1), ylim = c(1,0), col = c("white", "black"),
 main = "Original Image", asp = 1)
image(t(image2), xlim = c(0,1), ylim = c(1,0), col = c("white", "black"),
 main = "Skeleton Image", asp = 1)
)
}
}
\author{
Alexandre Thiery
}
\references{
F.W.M. Stentiford, R.G. Mortimer (1983), "Some new heuristics for thinning binary handprinted characters for OCR." \emph{IEEE Transactions on Systems, Man and Cybernetics}, \bold{13}, pp. 81-84.
}
\keyword{internal}

