% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MultivariateGamma}
\alias{MultivariateGamma}
\title{The multivariate gamma function.}
\usage{
MultivariateGamma(a, p, log = TRUE)
}
\arguments{
\item{a}{a non-negative numeric.}

\item{p}{a non-negative numeric.}

\item{log}{a logical value. If \code{TRUE}, multivariate gamma function is given as \code{log}. Default is \code{log=TRUE}.}
}
\value{
A numeric value.
}
\description{
Compute background parameters (overall and group means, within- and between group variances and covariances matrices) for the two level likelihood ratio.
}
\author{
Alexandre Thiery
}
\keyword{internal}

