% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TwoLevelLR}
\alias{TwoLevelLR}
\title{Two Level Multivariate Likelihood ratio}
\usage{
TwoLevelLR(data1, data2, background, n.iter = 1100, n.burnin = 100, nw)
}
\arguments{
\item{data1}{measurements from the 'reference' material.}

\item{data2}{measurements from the 'questioned' material.}

\item{background}{background parameters for the overall and group means, within- and between group variances and covariances matrice.}

\item{n.iter}{number of MCMC iterations. Default is \code{10000}.}

\item{n.burnin}{number of burn-in iterations. Default is \code{1000}.}

\item{nw}{degrees of freedom for the inverse Wishart distribution. Considering p variables in the data, nw must be > 2*p+4.}
}
\value{
The value of the log likelihood ratio (\eqn{log(LR)}).
}
\description{
Computes a two level multivariate likelihood ratio.
}
\details{
This methodology has been developed by \emph{Bozza et al. (2008)} for the assesment of evidence through the derivation of a likelihood ratio for multivariate data. It allows to take into account the correlation between variables, and the non-constant variability within sources.

In the context of handwritten expertise suppose that: (i) an anonymous letter (\emph{i.e.} the \emph{questioned document}) is available for comparative analysis, and (ii) written material from a suspect is selected for comparative purposes (\emph{i.e.} the \emph{reference document}. For the compuation of the likelihood ratio, we consider the following propositions of interest:
\itemize{
 \item{\eqn{H_p}}{the suspect is the author of the questioned document;}
 \item{\eqn{H_d}}{the suspect is not the author of the questioned document - a random person wrote the document.}
}
}
\examples{
\dontrun{
## Example where Hp is true.
## That is, when the suspect is the author of the questioned document

## DATASET: We use the `characterO` dataset. It contains the extracted Fourier (`n.fourier = 4`) parameters from 554 handwritten character loops, written by 11 writers. For more information, see `?characterO`.
data(characterO)

## PARAMETERS:
# data1: (reference document), the first 23 characters of writer 1
# data1: (questioned document), the last 23 characters of writer 1
# background: data from the remaining 10 writers
# n.iter: 11000
# n.burnin: 1000
# nw: 50


# data1, data2
data_reference = subset(characterO$measurements[,-1], subset = (characterO$info$writer == 1))[1:23,]
data_questioned = subset(characterO$measurements[,-1], subset = (characterO$info$writer == 1))[-(1:23),]

# background
subset = characterO$info$writer != 1
data_back = subset(characterO$measurements[,-1], subset = subset)
background = TwoLevelLR_Background(data_back, fac = as.factor(characterO$info$writer[subset]))

# others
n.iter = 11000
n.burnin = 1000
nw  = 50

# compute LLR
LLR = TwoLevelLR(data1 = data_reference,
                 data2 = data_questioned,
                 background = background,
                 n.iter = n.iter, n.burnin = n.burnin,
                 nw = nw)
LLR
}
}
\author{
Silvia Bozza \cr Alexandre Thiery
}
\references{
Bozza S, Taroni F, Marquis R and Schmittbuhl M (2008). "Probabilistic evaluation of handwriting evidence: likelihood ratio for authorship." \emph{Journal of the Royal Statistical Society: Series C (Applied Statistics)}, \strong{57} (3), pp. 329-341.
}
\seealso{
\code{\link{TwoLevelLR_Background}}
}

