% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TwoLevelLR_func}
\alias{TwoLevelLR_denominator}
\alias{TwoLevelLR_func}
\alias{TwoLevelLR_numerator}
\title{Functions for the Two Level Multivariate Likelihood ratio numerator and denominator.}
\usage{
TwoLevelLR_denominator(data.1, data.2, p, nw, n.iter, n.burnin, mu, B, logdetB,
  invB, invBmu, U, logdetU, IW.1, IW.2)

TwoLevelLR_numerator(data, p, nw, n.iter, n.burnin, mu, B, logdetB, invB,
  invBmu, U, logdetU, IW)
}
\description{
Functions for the Two Level Multivariate Likelihood ratio numerator and denominator.
}
\author{
Silvia Bozza \cr Alexandre Thiery
}
\references{
Bozza S, Taroni F, Marquis R and Schmittbuhl M (2008). "Probabilistic evaluation of handwriting evidence: likelihood ratio for authorship." \emph{Journal of the Royal Statistical Society: Series C (Applied Statistics)}, \strong{57} (3), pp. 329-341.
}
\seealso{
\code{\link{TwoLevelLR}}
}
\keyword{internal}

